rebase aws code
